Move decorators to proper app
