+++ landing_page_checker/views.py
@@ -6,7 +6,7 @@ from django.utils.decorators import method_decorator
 from django.views.generic import DetailView, ListView, UpdateView
 from django_otp.decorators import otp_required
 
-from common.decorators import directory_management_required
+from directory.decorators import directory_management_required
 from directory.models.settings import DirectorySettings
 from directory.models import DirectoryPage, SecuredropPage
 from landing_page_checker.forms import SecuredropPageOwnerForm
